+++ admin/src/test/java/org/bahmni/module/admin/csv/service/CSVRelationshipServiceTest.java
@@ -1,7 +1,7 @@
 package org.bahmni.module.admin.csv.service;
 
 import org.bahmni.module.admin.csv.models.RelationshipRow;
-import org.bahmni.module.bahmnicommons.service.BahmniPatientService;
+import org.bahmni.module.bahmnicommons.api.service.BahmniPatientService;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
