+++ bahmnicore-omod/src/test/java/org/bahmni/module/bahmnicore/web/v1_0/controller/BahmniVisitLocationControllerIT.java
@@ -4,7 +4,7 @@ package org.bahmni.module.bahmnicore.web.v1_0.controller;
 import org.bahmni.module.bahmnicore.web.v1_0.BaseIntegrationTest;
 import org.junit.Before;
 import org.junit.Test;
-import org.bahmni.module.bahmnicommons.visitlocation.VisitLocationNotFoundException;
+import org.bahmni.module.bahmnicommons.api.visitlocation.VisitLocationNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import static org.junit.Assert.assertEquals;
